+++ pandas_market_calendars/calendars/cme_globex_equities.py
@@ -1,6 +1,5 @@
-from .cme_globex_base import CMEGlobexBaseExchangeCalendar
-
 from datetime import time
+
 from pandas.tseries.holiday import AbstractHolidayCalendar
 from pytz import timezone
 
@@ -10,11 +9,12 @@ from pandas_market_calendars.holidays.cme import (
     USPresidentsDayBefore2015,
     USPresidentsDayAfter2015,
     GoodFridayBefore2021NotEarlyClose,
-    GoodFridayAfter2021,
     GoodFriday2010,
     GoodFriday2012,
     GoodFriday2015,
     GoodFriday2021,
+    GoodFriday2022,
+    GoodFridayAfter2022,
     USMemorialDay2013AndPrior,
     USMemorialDayAfter2013,
     USIndependenceDayBefore2022PreviousDay,
@@ -32,6 +32,7 @@ from pandas_market_calendars.holidays.us import (
     Christmas,
     USJuneteenthAfter2022,
 )
+from .cme_globex_base import CMEGlobexBaseExchangeCalendar
 
 
 class CMEGlobexEquitiesExchangeCalendar(CMEGlobexBaseExchangeCalendar):
@@ -61,7 +62,7 @@ class CMEGlobexEquitiesExchangeCalendar(CMEGlobexBaseExchangeCalendar):
             rules=[
                 USNewYearsDay,
                 GoodFridayBefore2021NotEarlyClose,
-                GoodFridayAfter2021,
+                GoodFriday2022,
                 Christmas,
             ]
         )
@@ -115,6 +116,7 @@ class CMEGlobexEquitiesExchangeCalendar(CMEGlobexBaseExchangeCalendar):
                         GoodFriday2012,
                         GoodFriday2015,
                         GoodFriday2021,
+                        GoodFridayAfter2022,
                     ]
                 ),
             ),