+++ pandas_market_calendars/calendars/jpx.py
@@ -24,7 +24,7 @@ class JPXExchangeCalendar(MarketCalendar):
     Close Time: 4:00 PM, Asia/Tokyo
     """
 
-    aliases = ["JPX"]
+    aliases = ["JPX", "XJPX"]
     regular_market_times = {
         "market_open": ((None, time(9)),),
         "market_close": ((None, time(15)),),