+++ pandas_market_calendars/calendars/tase.py
@@ -1,6 +1,8 @@
 from datetime import time
+
 from pandas import Timestamp
 from pytz import timezone
+
 from pandas_market_calendars.market_calendar import MarketCalendar
 
 TASEClosedDay = [