+++ tests/test_bse_calendar.py
@@ -1,6 +1,7 @@
 import datetime
-import pytest
+
 import pandas as pd
+import pytest
 import pytz
 
 from pandas_market_calendars.calendars.bse import BSEExchangeCalendar, BSEClosedDay