+++ tests/test_exchange_calendar_cme_globex_crypto.py
@@ -1,4 +1,5 @@
 import datetime as dt
+
 import pandas as pd
 import pytest
 from pandas.tseries.offsets import Day, Hour, Minute
@@ -260,6 +261,7 @@ def test_is_different():
         ("2022-12-30", "open"),
         ("2023-01-02", "closed"),
         ("2023-01-03", "open"),
+        ("2023-04-07", "1015"),
     ],
     ids=lambda x: f"{x[0]} {x[1]}",
 )
@@ -270,7 +272,7 @@ def test_2020_through_2022_and_prior_holidays(day_status):
 
     year = int(day_str.split("-")[0])
     under_test = CMEGlobexCryptoExchangeCalendar()
-    schedule = under_test.schedule(f"{year}-01-01", f"{year+1}-01-01", tz=TZ)
+    schedule = under_test.schedule(f"{year}-01-01", f"{year + 1}-01-01", tz=TZ)
 
     if expected_status == "open":
         s = schedule.loc[day_str]