+++ tests/test_exchange_calendar_cme_globex_energy_and_metals.py
@@ -1,13 +1,10 @@
-import pytz
-
 import pandas as pd
+import pytz
 from pandas.testing import assert_index_equal
-from pandas.tseries.offsets import CustomBusinessDay
 
 from pandas_market_calendars.calendars.cme_globex_energy_and_metals import (
     CMEGlobexEnergyAndMetalsExchangeCalendar,
 )
-
 
 cal = CMEGlobexEnergyAndMetalsExchangeCalendar()
 
@@ -68,7 +65,7 @@ def _test_has_late_opens(late_opens, start, end):
     expected = cal.late_opens(schedule)
     assert len(expected) == len(late_opens)
     for ts in late_opens:
-        assert _test_verify_late_open_time(schedule, ts) == True
+        assert _test_verify_late_open_time(schedule, ts) is True
 
 
 def _test_verify_early_close_time(schedule, timestamp):
@@ -84,7 +81,7 @@ def _test_has_early_closes(early_closes, start, end):
     expected = cal.early_closes(schedule)
     assert len(expected) == len(early_closes)
     for ts in early_closes:
-        assert _test_verify_early_close_time(schedule, ts) == True
+        assert _test_verify_early_close_time(schedule, ts) is True
 
 
 #########################################################################