+++ tests/test_exchange_calendar_cme_globex_fx.py
@@ -143,6 +143,7 @@ def test_sunday_opens():
         ("2022-12-30", "open"),
         ("2023-01-02", "closed"),
         ("2023-01-03", "open"),
+        ("2023-04-07", "1015"),
     ],
     ids=lambda x: f"{x[0]} {x[1]}",
 )
@@ -152,7 +153,7 @@ def test_2020_through_2022_and_prior_holidays(day_status):
     expected_status = day_status[1]
 
     under_test = CMEGlobexFXExchangeCalendar()
-    schedule = under_test.schedule("2020-01-01", "2023-02-28", tz=TZ)
+    schedule = under_test.schedule("2020-01-01", "2023-04-28", tz=TZ)
 
     if expected_status == "open":
         s = schedule.loc[day_str]